requirements/requirements.txt
rasterio==1.3.9
numpy==1.23.0
scikit-learn==1.3.0
scikit-image==0.21.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0
keras== 2.12.0
pillow==10.4.0
geopandas==0.13.2
shapely==2.0.2
opencv-python==4.7.0.72
tqdm== 4.66.4
rtree==1.3.0
utm==0.7.0
ultralytics==8.2.52
supervision==0.22.0
geojson==3.1.0
